# openapi.yaml
openapi: 3.1.0
info:
  title: LearnFast Core Engine
  description: Hybrid Graph-RAG Learning Platform API
  version: 0.1.0
paths:
...
  /api/ai/generate-flashcards:
    post:
      tags:
      - AI Generation
      summary: Generate Flashcards
      description: Generates flashcards from a document.
      operationId: generate_flashcards_api_ai_generate_flashcards_post
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/GenerateRequest'
        required: true
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema: {}
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /api/ai/generate-questions:
    post:
      tags:
      - AI Generation
      summary: Generate Questions
      description: Generates multiple-choice questions from a document.
      operationId: generate_questions_api_ai_generate_questions_post
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/GenerateRequest'
        required: true
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema: {}
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /api/ai/learning-path:
    post:
      tags:
      - AI Generation
      summary: Generate Learning Path
      description: "Generates a structured learning path/curriculum.\n\nAttempts to\
        \ resolve path using Knowledge Graph first. \nFalls back to LLM generation\
        \ if no graph path found or if document context provided."
      operationId: generate_learning_path_api_ai_learning_path_post
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/PathRequest'
        required: true
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema: {}
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
...
components:
  schemas:
...
    GenerateRequest:
      properties:
        document_id:
          type: integer
          title: Document Id
        count:
          type: integer
          title: Count
          default: 5
        llm_config:
          anyOf:
          - $ref: '#/components/schemas/LLMConfig'
          - type: 'null'
      type: object
      required:
      - document_id
      title: GenerateRequest
      description: Request schema for flashcard/question generation.
    HTTPValidationError:
      properties:
        detail:
          items:
            $ref: '#/components/schemas/ValidationError'
          type: array
          title: Detail
      type: object
      title: HTTPValidationError
    LLMConfig:
      properties:
        provider:
          type: string
          title: Provider
          default: openai
        api_key:
          anyOf:
          - type: string
          - type: 'null'
          title: Api Key
        base_url:
          anyOf:
          - type: string
          - type: 'null'
          title: Base Url
        model:
          anyOf:
          - type: string
          - type: 'null'
          title: Model
      type: object
      title: LLMConfig
      description: Configuration overrides for LLM generation.
    PathRequest:
      properties:
        user_id:
          type: string
          title: User Id
        target_concept:
          type: string
          title: Target Concept
        time_budget_minutes:
          type: integer
          title: Time Budget Minutes
          default: 60
        document_id:
          anyOf:
          - type: integer
          - type: 'null'
          title: Document Id
      type: object
      required:
      - user_id
      - target_concept
      title: PathRequest
      description: Request for generating a learning path.
...
    ValidationError:
      properties:
        loc:
          items:
            anyOf:
            - type: string
            - type: integer
          type: array
          title: Location
        msg:
          type: string
          title: Message
        type:
          type: string
          title: Error Type
      type: object
      required:
      - loc
      - msg
      - type
      title: ValidationError